requirements.txt
openai~=1.51.0
requests~=2.32.3
google-generativeai~=0.7.2
google~=3.0.0
anthropic~=0.35.0
pillow~=10.4.0